+++ src/CG.Core/Security/ClaimsPrincipalExtensions.cs
@@ -31,7 +31,7 @@ namespace System.Security.Claims
 
             // Look for the claim on the principal.
             var claim = principal.Claims.FirstOrDefault(
-                x => x.Type == "email"
+                x => x.Type == ClaimTypes.Email
                 );
             
             // Did we find it?
@@ -43,6 +43,122 @@ namespace System.Security.Claims
 
             // Claim not found.
             return string.Empty;
+        }
+
+        // *******************************************************************
+
+        /// <summary>
+        /// This method returns the value of the name identifier claim, if it 
+        /// exists, in the specified <see cref="ClaimsPrincipal"/> object.
+        /// </summary>
+        /// <param name="claimsPrincipal">The principal to use for the oepration.</param>
+        /// <returns>The value of the claim, or an empty string, if the claim
+        /// wasn't found on the pricipal.</returns>
+        public static string GetNameIdentifier(
+            this ClaimsPrincipal claimsPrincipal
+            )
+        {
+            // Look for the specified claim.
+            var claim = claimsPrincipal.Claims.FirstOrDefault(
+                x => x.Type == ClaimTypes.NameIdentifier
+                );
+
+            // Did we find it?
+            if (null != claim)
+            {
+                return claim.Value;
+            }
+            else
+            {
+                return String.Empty; // Didn't find it.
+            }
+        }
+
+        // *******************************************************************
+
+        /// <summary>
+        /// This method returns the value of the nickname claim, if it 
+        /// exists, in the specified <see cref="ClaimsPrincipal"/> object.
+        /// </summary>
+        /// <param name="claimsPrincipal">The principal to use for the oepration.</param>
+        /// <returns>The value of the claim, or an empty string, if the claim
+        /// wasn't found on the pricipal.</returns>
+        public static string GetNickName(
+            this ClaimsPrincipal claimsPrincipal
+            )
+        {
+            // Look for the specified claim.
+            var claim = claimsPrincipal.Claims.FirstOrDefault(
+                x => x.Type == "nickname"
+                );
+
+            // Did we find it?
+            if (null != claim)
+            {
+                return claim.Value;
+            }
+            else
+            {
+                return String.Empty; // Didn't find it.
+            }
+        }
+
+        // *******************************************************************
+
+        /// <summary>
+        /// This method returns the value of the name claim, if it 
+        /// exists, in the specified <see cref="ClaimsPrincipal"/> object.
+        /// </summary>
+        /// <param name="claimsPrincipal">The principal to use for the oepration.</param>
+        /// <returns>The value of the claim, or an empty string, if the claim
+        /// wasn't found on the pricipal.</returns>
+        public static string GetName(
+            this ClaimsPrincipal claimsPrincipal
+            )
+        {
+            // Look for the specified claim.
+            var claim = claimsPrincipal.Claims.FirstOrDefault(
+                x => x.Type == ClaimTypes.Name
+                );
+
+            // Did we find it?
+            if (null != claim)
+            {
+                return claim.Value;
+            }
+            else
+            {
+                return String.Empty; // Didn't find it.
+            }
+        }
+
+        // *******************************************************************
+
+        /// <summary>
+        /// This method returns the value of the picture claim, if it 
+        /// exists, in the specified <see cref="ClaimsPrincipal"/> object.
+        /// </summary>
+        /// <param name="claimsPrincipal">The principal to use for the oepration.</param>
+        /// <returns>The value of the claim, or an empty string, if the claim
+        /// wasn't found on the pricipal.</returns>
+        public static string GetPicture(
+            this ClaimsPrincipal claimsPrincipal
+            )
+        {
+            // Look for the specified claim.
+            var claim = claimsPrincipal.Claims.FirstOrDefault(
+                x => x.Type == "picture"
+                );
+
+            // Did we find it?
+            if (null != claim)
+            {
+                return claim.Value;
+            }
+            else
+            {
+                return String.Empty; // Didn't find it.
+            }
         }
 
         #endregion